package/autoretriever/requirements.txt
prompt_toolkit
openai==1.3.5
unstructured[all-docs]
pyautogen[retrievechat,teachable,mathchat]==0.2.0
chromadb